Integrate Volley builds with errorprone. (#162)
All error prone warnings+errors now lead to compile failures.
Existing failures have been fixed or suppressed.

Fixes #100

+++ src/main/java/com/android/volley/DefaultRetryPolicy.java
@@ -90,7 +90,7 @@ public class DefaultRetryPolicy implements RetryPolicy {
     @Override
     public void retry(VolleyError error) throws VolleyError {
         mCurrentRetryCount++;
-        mCurrentTimeoutMs += (mCurrentTimeoutMs * mBackoffMultiplier);
+        mCurrentTimeoutMs += (int) (mCurrentTimeoutMs * mBackoffMultiplier);
         if (!hasAttemptRemaining()) {
             throw error;
         }
